+++ src/main/java/racingcar/domain/Cars.java
@@ -1,8 +1,10 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.CarDto;
+import racingcar.utils.Util;
 
 public class Cars {
