+++ src/main/java/racingcar/domain/Position.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class Position {
