+++ src/main/java/racingcar/utils/Util.java
@@ -1,9 +1,11 @@
-package racingcar;
+package racingcar.utils;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.domain.Car;
+import racingcar.CarDto;
 
 public class Util {
