+++ src/main/java/racingcar/utils/Validator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.utils;
 
 public class Validator {
