+++ src/main/java/racingcar/view/InputView.java
@@ -1,6 +1,7 @@
-package racingcar;
+package racingcar.view;
 
 import camp.nextstep.edu.missionutils.Console;
+import racingcar.utils.Validator;
 
 public class InputView {
